requirements.txt
fastapi
uvicorn
sarvamai
pydub
python-dotenv
